requirements.txt
audioread==3.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
decord==0.6.0
diskcache==5.6.3
easyocr==1.7.2
faiss-cpu==1.12.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
#llama_cpp_python @ file:///home/tamta/memory/llama_cpp_python-0.3.16-cp310-cp310-linux_x86_64.whl#sha256=d1ec4c273dad11f08bde6ddb1728f6897c2bc6ab5004fb2ed12290c68ef4ccfd
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
proglog==0.1.12
pyclipper==1.3.0.post6
pycparser==2.23
pyparsing==3.2.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scenedetect==0.6.7.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
shapely==2.1.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
#torch==2.6.0
#torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14